Split pkg/utils/kubernetes/health/health.go into separate files
