+++ pkg/utils/kubernetes/health/deployment.go
@@ -1,0 +1,87 @@
+// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package health
+
+import (
+	"fmt"
+
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func getDeploymentCondition(conditions []appsv1.DeploymentCondition, conditionType appsv1.DeploymentConditionType) *appsv1.DeploymentCondition {
+	for _, condition := range conditions {
+		if condition.Type == conditionType {
+			return &condition
+		}
+	}
+	return nil
+}
+
+var (
+	trueDeploymentConditionTypes = []appsv1.DeploymentConditionType{
+		appsv1.DeploymentAvailable,
+	}
+
+	trueOptionalDeploymentConditionTypes = []appsv1.DeploymentConditionType{
+		appsv1.DeploymentProgressing,
+	}
+
+	falseOptionalDeploymentConditionTypes = []appsv1.DeploymentConditionType{
+		appsv1.DeploymentReplicaFailure,
+	}
+)
+
+// CheckDeployment checks whether the given Deployment is healthy.
+// A deployment is considered healthy if the controller observed its current revision and
+// if the number of updated replicas is equal to the number of replicas.
+func CheckDeployment(deployment *appsv1.Deployment) error {
+	if deployment.Status.ObservedGeneration < deployment.Generation {
+		return fmt.Errorf("observed generation outdated (%d/%d)", deployment.Status.ObservedGeneration, deployment.Generation)
+	}
+
+	for _, trueConditionType := range trueDeploymentConditionTypes {
+		conditionType := string(trueConditionType)
+		condition := getDeploymentCondition(deployment.Status.Conditions, trueConditionType)
+		if condition == nil {
+			return requiredConditionMissing(conditionType)
+		}
+		if err := checkConditionState(string(corev1.ConditionTrue), string(condition.Status), condition.Reason, condition.Message); err != nil {
+			return err
+		}
+	}
+
+	for _, trueOptionalConditionType := range trueOptionalDeploymentConditionTypes {
+		condition := getDeploymentCondition(deployment.Status.Conditions, trueOptionalConditionType)
+		if condition == nil {
+			continue
+		}
+		if err := checkConditionState(string(corev1.ConditionTrue), string(condition.Status), condition.Reason, condition.Message); err != nil {
+			return err
+		}
+	}
+
+	for _, falseOptionalConditionType := range falseOptionalDeploymentConditionTypes {
+		condition := getDeploymentCondition(deployment.Status.Conditions, falseOptionalConditionType)
+		if condition == nil {
+			continue
+		}
+		if err := checkConditionState(string(corev1.ConditionFalse), string(condition.Status), condition.Reason, condition.Message); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
