+++ pkg/utils/kubernetes/health/deployment_test.go
@@ -1,0 +1,86 @@
+// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package health_test
+
+import (
+	"github.com/gardener/gardener/pkg/utils/kubernetes/health"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/extensions/table"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/types"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+var _ = Describe("Deployment", func() {
+	Describe("CheckDeployment", func() {
+		DescribeTable("deployments",
+			func(deployment *appsv1.Deployment, matcher types.GomegaMatcher) {
+				err := health.CheckDeployment(deployment)
+				Expect(err).To(matcher)
+			},
+			Entry("healthy", &appsv1.Deployment{
+				Status: appsv1.DeploymentStatus{Conditions: []appsv1.DeploymentCondition{
+					{
+						Type:   appsv1.DeploymentAvailable,
+						Status: corev1.ConditionTrue,
+					},
+				}},
+			}, BeNil()),
+			Entry("healthy with progressing", &appsv1.Deployment{
+				Status: appsv1.DeploymentStatus{Conditions: []appsv1.DeploymentCondition{
+					{
+						Type:   appsv1.DeploymentAvailable,
+						Status: corev1.ConditionTrue,
+					},
+					{
+						Type:   appsv1.DeploymentProgressing,
+						Status: corev1.ConditionTrue,
+					},
+				}},
+			}, BeNil()),
+			Entry("not observed at latest version", &appsv1.Deployment{
+				ObjectMeta: metav1.ObjectMeta{Generation: 1},
+			}, HaveOccurred()),
+			Entry("not available", &appsv1.Deployment{
+				Status: appsv1.DeploymentStatus{Conditions: []appsv1.DeploymentCondition{
+					{
+						Type:   appsv1.DeploymentAvailable,
+						Status: corev1.ConditionFalse,
+					},
+					{
+						Type:   appsv1.DeploymentProgressing,
+						Status: corev1.ConditionTrue,
+					},
+				}},
+			}, HaveOccurred()),
+			Entry("not progressing", &appsv1.Deployment{
+				Status: appsv1.DeploymentStatus{Conditions: []appsv1.DeploymentCondition{
+					{
+						Type:   appsv1.DeploymentAvailable,
+						Status: corev1.ConditionTrue,
+					},
+					{
+						Type:   appsv1.DeploymentProgressing,
+						Status: corev1.ConditionFalse,
+					},
+				}},
+			}, HaveOccurred()),
+			Entry("available | progressing missing", &appsv1.Deployment{}, HaveOccurred()),
+		)
+	})
+})
