+++ pkg/utils/kubernetes/health/etcd.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package health
+
+import (
+	"fmt"
+
+	"github.com/gardener/gardener/pkg/utils"
+
+	druidv1alpha1 "github.com/gardener/etcd-druid/api/v1alpha1"
+)
+
+// CheckEtcd checks whether the given Etcd is healthy.
+// A Etcd is considered healthy if its ready field in status is true.
+func CheckEtcd(etcd *druidv1alpha1.Etcd) error {
+	if !utils.IsTrue(etcd.Status.Ready) {
+		return fmt.Errorf("etcd %q is not ready yet", etcd.Name)
+	}
+	return nil
+}
