+++ pkg/utils/kubernetes/health/node_test.go
@@ -1,0 +1,43 @@
+// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package health_test
+
+import (
+	"github.com/gardener/gardener/pkg/utils/kubernetes/health"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/extensions/table"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/types"
+	corev1 "k8s.io/api/core/v1"
+)
+
+var _ = Describe("Node", func() {
+	Describe("CheckNode", func() {
+		DescribeTable("nodes",
+			func(node *corev1.Node, matcher types.GomegaMatcher) {
+				err := health.CheckNode(node)
+				Expect(err).To(matcher)
+			},
+			Entry("healthy", &corev1.Node{
+				Status: corev1.NodeStatus{Conditions: []corev1.NodeCondition{{Type: corev1.NodeReady, Status: corev1.ConditionTrue}}},
+			}, BeNil()),
+			Entry("no ready condition", &corev1.Node{}, HaveOccurred()),
+			Entry("ready condition not indicating true", &corev1.Node{
+				Status: corev1.NodeStatus{Conditions: []corev1.NodeCondition{{Type: corev1.NodeReady, Status: corev1.ConditionFalse}}},
+			}, HaveOccurred()),
+		)
+	})
+})
